unit of work pattern section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MembersService } from '../_services/members.service';
 import { MessageService } from '../_services/message.service';
 
@@ -17,7 +18,9 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService, private memberService: MembersService) { }
+  constructor(private messageService: MessageService, 
+    private memberService: MembersService,
+    private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -40,8 +43,12 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.memberService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Are you sure about deleting the message', 'This cannot be undone!').subscribe((result) => {
+      if(result) {
+        this.memberService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
   }
